requirements.txt
argcomplete==2.0.0
asgiref==3.6.0
boto3==1.26.64
botocore==1.29.64
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.6
django-filter==22.1
django-s3-sqlite==0.0.5
django-s3-storage==0.13.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
durationpy==0.5
Faker==16.6.1
hjson==3.1.0
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.2
packaging==23.0
placebo==0.9.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-slugify==8.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.3.0
tzdata==2022.7
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.2
zappa==0.56.1
zappa-django-utils==0.4.1
